API: Surface all remote config options
- Combined serializers for /remotes and /sync/config as the model is the
  same
- Used a Mixin to deduplicate existing code
- Added extra fields to the serializer: username, password, proxy_url,
  client_key, client_cert, tls_validation, ca_cert, pulp_href, pk,
download_concurrecy

Issue: AAH-21
