+++ galaxy_ng/app/api/v3/viewsets/sync.py
@@ -3,7 +3,7 @@ from django.http import Http404
 from pulp_ansible.app.models import AnsibleDistribution
 from galaxy_ng.app.access_control import access_policy
 from galaxy_ng.app.api import base as api_base
-from ..serializers.sync import SyncConfigSerializer
+from ..serializers.sync import CollectionRemoteSerializer
 
 
 class SyncConfigViewSet(
@@ -11,7 +11,7 @@ class SyncConfigViewSet(
     mixins.UpdateModelMixin,
     api_base.GenericViewSet,
 ):
-    serializer_class = SyncConfigSerializer
+    serializer_class = CollectionRemoteSerializer
     permission_classes = [access_policy.CollectionRemoteAccessPolicy]
 
     def get_object(self):
